tools: add conditionals in mk-sdcard-fs-tarball script
This script supports only the beaglebone product and
should not be run anywhere else as-is.

+++ tools/mk-sdcard-fs-tarball.sh
@@ -6,12 +6,32 @@
 
 # Based on rowboat top-level Makefile
 
+# first check if lunch is done
+if [ -z "$ANDROID_BUILD_TOP" ]; then
+    echo "ANDROID_BUILD_TOP undefined"
+exit 1
+fi
+if [ -z "$ANDROID_PRODUCT_OUT" ]; then
+    echo "ANDROID_PRODUCT_OUT undefined"
+exit 1
+fi
+
+# Check for beaglebone target
+if [ "$TARGET_PRODUCT" != "beaglebone" ]; then
+    echo "Only 'TARGET_PRODUCT = beaglebone' is supported for generating rootfs tarball"
+exit 1
+fi
+
 export ANDROID_FS_DIR=$ANDROID_PRODUCT_OUT/android_rootfs
 
-# TODO: Add conditional
 echo "ANDROID_FS_DIR: $ANDROID_FS_DIR"
-rm -rf $ANDROID_FS_DIR
-rm $ANDROID_BUILD_TOP/rootfs.tar.bz2
+if [ -d "$ANDROID_FS_DIR" ]; then
+    echo "Deleting old fs tarball"
+    rm -rf $ANDROID_FS_DIR
+fi
+if [ -e "$ANDROID_BUILD_TOP/rootfs.tar.bz2" ]; then
+    rm $ANDROID_BUILD_TOP/rootfs.tar.bz2
+fi
 mkdir $ANDROID_FS_DIR
 
 echo "Copying files..."
